Just put every method in it's file
