+++ api/admin/read.php
@@ -1,0 +1,22 @@
+<?php
+
+session_start();
+require_once(__DIR__ . "/../../controller/Admin.php");
+require_once(__DIR__ . "/../../model/errors.php");
+require_once(__DIR__ . "/../../config.php");
+
+use controller\Admin;
+
+header("Access-Control-Allow-Origin:*");
+header("Access-Control-Allow-Methods:DELETE");
+header("Access-Control-Allow-Headers:Content-Type,Authorization,X-Request-With");
+
+/**@var Admin $admin*/
+$admin = unserialize($_SESSION['user']);
+if (!$admin) {
+  error422("Unauthorized", 401);
+}
+
+$data = getData();
+
+$admin->read($data['id']);
